ItemEnhancement cutoff mail
